requirements.txt
pygame==2.4.0
PyOpenGL==3.1.7
